refactors dosplay product components
