+++ frontend/src/App.tsx
@@ -1,22 +1,24 @@
-import PrimarySearchAppBar from "./components/navbar/Header";
-import DisplayProducts from "./components/display-products/DisplayProducts";
-import { BrowserRouter, Route } from "react-router-dom";
-import { Routes } from "react-router";
-import DisplayProduct from "./components/display-product/DisplayProduct";
-import { QueryClientProvider, QueryClient } from "react-query"
-import Footer from "./components/footer/Footer";
-import { ShoppingCartProvider } from "./contexts/shopping-cart-context/shoppingCartContext";
-import CartPage from "./components/cart-page";
-import SignIn from "./components/authentication/sign-in";
-import ShippingAddressUi from "./components/checkout/shopping-address";
-import { ToastContainer } from "react-toastify"
-import 'react-toastify/dist/ReactToastify.css'
-import SignUp from "./components/authentication/sign-up/index";
-import PaymentUi from "./components/checkout/PaymentUi";
-import PlaceOrderUi from "./components/checkout/place-order";
-import OrderScreen from "./components/orders/OrderScreen";
-import OrdersHistory from "./components/orders/order-history";
-import ProfileUpdate from "./components/ProfileUpdate";
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { QueryClient, QueryClientProvider } from 'react-query';
+import { ToastContainer } from 'react-toastify';
+
+import PrimarySearchAppBar from './components/navbar/Header';
+import DisplayProducts from './components/display-products/DisplayProducts';
+import DisplayProduct from './components/display-product';
+import CartPage from './components/cart-page';
+import SignIn from './components/authentication/sign-in';
+import ShippingAddressUi from './components/checkout/shopping-address';
+import SignUp from './components/authentication/sign-up/index';
+import PaymentUi from './components/checkout/PaymentUi';
+import PlaceOrderUi from './components/checkout/place-order';
+import OrderScreen from './components/orders/OrderScreen';
+import OrdersHistory from './components/orders/order-history';
+import ProfileUpdate from './components/ProfileUpdate';
+import Footer from './components/footer/Footer';
+import { ShoppingCartProvider } from './contexts/shopping-cart-context/shoppingCartContext';
+
+import 'react-toastify/dist/ReactToastify.css';
+
 const reactQueryClient = new QueryClient()
 
 function App() {
