+++ frontend/src/components/display-product/index.tsx
